FOV 120; minimumCost 250; Kostenfunktion noch nicht ausgereift

+++ src/visualize.h
@@ -17,7 +17,7 @@ class Visualizer{
         // Visualizer(double scale_x, double scale_y, double scale_z);
         Visualizer();
         ~Visualizer();
-        void setMarkerArray(ros::NodeHandle &nh, std::vector<geometry_msgs::Pose> vizPos, int r, int g, int b, bool deleteArray);
+        void setMarkerArray(ros::NodeHandle &nh, std::vector<geometry_msgs::Pose> vizPos, double r, double g, double b, bool deleteArray);
         // void setMarker();
         void deleteMarkerArray(ros::NodeHandle &nh);
 };
